% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISM.R
\name{ISM}
\alias{ISM}
\title{Interpretive Structural Modeling (ISM).}
\usage{
ISM(fname, Dir)
}
\arguments{
\item{fname}{a matrix consists of 1s' and 0's (initial reachability matrix)}

\item{Dir}{a path where user wants to save output files}
}
\value{
provides two output files (Final Reachability Matrix and Level Partition of each iteration)
in Excel format
}
\description{
This methods provides a wellformated solution of ISM
}
\details{
This Function Provides well-formatted
and readable excel output files (Final Reachability Matrix and Level Partition of each iteration) that
make interpretation easier.
}
\examples{
ISM(fname=matrix(c(1,1,1,1,1,0,1,1,1,1,0,0,1,0,0,0,1,1,1,1,0,1,1,0,1),5,5,byrow=TRUE),Dir=tempdir())

}
\references{
Adarsh Anand, Gunjan Bansal, (2017) "Interpretive structural modelling for attributes of software quality", Journal of Advances in Management Research, Vol. 14 Issue: 3, pp.256-269, https://doi.org/10.1108/JAMR-11-2016-0097
}
\author{
Adarsh Anand, Gunjan Bansal
}
