% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISM.R
\name{Mat_format}
\alias{Mat_format}
\title{This Mat_format Function formats the ISM_Matrix.xlsx file That is implicitly called by ISM.}
\usage{
Mat_format(fin_mat, A_mat, file2)
}
\arguments{
\item{fin_mat}{a final matrix consists of 1s' and 0's (final reachability matrix) produced by \code{ISM}}

\item{A_mat}{a initial matrix consists of 1s' and 0's (initial reachability matrix) produced by \code{ISM}}

\item{file2}{a final matrix consists of 1s' and 0's (final reachability matrix) produced by \code{ISM}}
}
\description{
This Mat_format Function formats the ISM_Matrix.xlsx file That is implicitly called by ISM.
}
