% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISM.R
\name{outputformat}
\alias{outputformat}
\title{This outputformat Function formats the ISM_output.xlsx file that implicitly called by ISM.}
\usage{
outputformat(file1)
}
\arguments{
\item{file1}{a Level out iterations produced by \code{ISM}}
}
\description{
This outputformat Function formats the ISM_output.xlsx file that implicitly called by ISM.
}
